requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
jinja2
fpdf2
pydantic[email]